EES-3789 Move PublicationTitleOrSlugChanged to Admin's MethodologyService
