+++ src/GovUk.Education.ExploreEducationStatistics.Content.Services.Tests/MethodologyServiceTests.cs
@@ -9,10 +9,8 @@ using GovUk.Education.ExploreEducationStatistics.Common.Utils;
 using GovUk.Education.ExploreEducationStatistics.Content.Model;
 using GovUk.Education.ExploreEducationStatistics.Content.Model.Database;
 using GovUk.Education.ExploreEducationStatistics.Content.Model.Repository.Interfaces;
-using GovUk.Education.ExploreEducationStatistics.Content.Services.Interfaces.Cache;
 using GovUk.Education.ExploreEducationStatistics.Content.Services.Tests.Mappings;
 using GovUk.Education.ExploreEducationStatistics.Content.Services.ViewModels;
-using Microsoft.EntityFrameworkCore;
 using Moq;
 using Xunit;
 using static GovUk.Education.ExploreEducationStatistics.Common.Services.CollectionUtils;
@@ -695,569 +693,16 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Services.Tests
             VerifyAllMocks(methodologyVersionRepository);
         }
 
-        [Fact]
-        public async Task PublicationTitleOrSlugChanged()
-        {
-            var publicationId = Guid.NewGuid();
-            var originalVersionId = Guid.NewGuid();
-            var latestPublishedVersionId = Guid.NewGuid();
-
-            var contentDbContextId = Guid.NewGuid().ToString();
-            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
-            {
-                var publicationMethodology = new PublicationMethodology
-                {
-                    PublicationId = publicationId,
-                    Owner = true,
-                    Methodology = new Methodology
-                    {
-                        LatestPublishedVersionId = latestPublishedVersionId,
-                        OwningPublicationTitle = "Original Title",
-                        OwningPublicationSlug = "original-slug",
-                        Versions = new List<MethodologyVersion>
-                        {
-                            new()
-                            {
-                                Id = originalVersionId,
-                                Version = 0,
-                            },
-                            new()
-                            {
-                                Id = latestPublishedVersionId,
-                                Version = 1,
-                                PreviousVersionId = originalVersionId,
-                            },
-                        },
-                    },
-                };
-
-                await contentDbContext.PublicationMethodologies.AddAsync(publicationMethodology);
-                await contentDbContext.SaveChangesAsync();
-            }
-
-            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
-            {
-                var redirectsCacheService = new Mock<IRedirectsCacheService>(MockBehavior.Strict);
-                redirectsCacheService.Setup(mock => mock.UpdateRedirects())
-                    .ReturnsAsync(new RedirectsViewModel(new List<RedirectViewModel>(), new List<RedirectViewModel>()));
-
-                var service = SetupMethodologyService(contentDbContext,
-                    redirectsCacheService: redirectsCacheService.Object);
-                await service.PublicationTitleOrSlugChanged(publicationId, "original-slug", "New Title", "new-slug");
-            }
-
-            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
-            {
-                var publicationMethodology = await contentDbContext
-                    .PublicationMethodologies
-                    .Include(m => m.Methodology.Versions)
-                    .SingleAsync(m => m.PublicationId == publicationId);
-
-                Assert.Equal("New Title", publicationMethodology.Methodology.OwningPublicationTitle);
-                Assert.Equal("new-slug", publicationMethodology.Methodology.OwningPublicationSlug);
-
-                // As no AlternativeTitle or AlternativeSlug set, the MethodologyVersion's title and slug will also change
-                Assert.Equal("New Title", publicationMethodology.Methodology.Versions[1].Title);
-                Assert.Equal("new-slug", publicationMethodology.Methodology.Versions[1].Slug);
-
-                // As methodology is published and it's slug has changed, a redirect is created for LatestPublishedVersion
-                var methodologyRedirects = await contentDbContext.MethodologyRedirects
-                    .ToListAsync();
-                var methodologyRedirect = Assert.Single(methodologyRedirects);
-                Assert.Equal(latestPublishedVersionId, methodologyRedirect.MethodologyVersionId);
-                Assert.Equal("original-slug", methodologyRedirect.Slug);
-            }
-        }
-
-        [Fact]
-        public async Task PublicationTitleOrSlugChanged_NoMethodologyRedirectAsMethodologyUnpublished()
-        {
-            var publicationId = Guid.NewGuid();
-
-            var contentDbContextId = Guid.NewGuid().ToString();
-            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
-            {
-                var publicationMethodology = new PublicationMethodology
-                {
-                    PublicationId = publicationId,
-                    Owner = true,
-                    Methodology = new Methodology
-                    {
-                        LatestPublishedVersionId = null,
-                        OwningPublicationTitle = "Original Title",
-                        OwningPublicationSlug = "original-slug",
-                        Versions = ListOf(new MethodologyVersion()),
-                    },
-                };
-
-                await contentDbContext.PublicationMethodologies.AddAsync(publicationMethodology);
-                await contentDbContext.SaveChangesAsync();
-            }
-
-            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
-            {
-                var redirectsCacheService = new Mock<IRedirectsCacheService>(MockBehavior.Strict);
-                redirectsCacheService.Setup(mock => mock.UpdateRedirects())
-                    .ReturnsAsync(new RedirectsViewModel(new List<RedirectViewModel>(), new List<RedirectViewModel>()));
-
-                var service = SetupMethodologyService(contentDbContext,
-                    redirectsCacheService: redirectsCacheService.Object);
-                await service.PublicationTitleOrSlugChanged(publicationId,
-                    "original-slug", "New Title", "new-slug");
-            }
-
-            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
-            {
-                var publicationMethodology = await contentDbContext
-                    .PublicationMethodologies
-                    .Include(m => m.Methodology.Versions)
-                    .SingleAsync(m => m.PublicationId == publicationId);
-
-                Assert.Equal("New Title", publicationMethodology.Methodology.OwningPublicationTitle);
-                Assert.Equal("new-slug", publicationMethodology.Methodology.OwningPublicationSlug);
-
-                // As the Publication's Title and Slug changed, and no AlternateTitle/Slug set,
-                // the methodology's title and slug will also change
-                Assert.Equal("New Title", publicationMethodology.Methodology.Versions[0].Title);
-                Assert.Equal("new-slug", publicationMethodology.Methodology.Versions[0].Slug);
-
-                // Methodology is unpublished, so no redirect
-                var methodologyRedirects = await contentDbContext.MethodologyRedirects
-                    .ToListAsync();
-                Assert.Empty(methodologyRedirects);
-            }
-        }
-
-
-
-        [Fact]
-        public async Task PublicationTitleOrSlugChanged_DoesNotAffectUnrelatedMethodologies()
-        {
-            var publicationId = Guid.NewGuid();
-            var unrelatedPublicationId = Guid.NewGuid();
-
-            var contentDbContextId = Guid.NewGuid().ToString();
-
-            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
-            {
-                var publicationMethodology = new PublicationMethodology
-                {
-                    PublicationId = publicationId,
-                    Owner = true,
-                    Methodology = new Methodology
-                    {
-                        Versions = ListOf(new MethodologyVersion
-                        {
-                            Status = Draft
-                        }),
-                        OwningPublicationTitle = "Original Title",
-                        OwningPublicationSlug = "original-slug",
-                    }
-                };
-
-                var unrelatedPublicationMethodology = new PublicationMethodology
-                {
-                    PublicationId = unrelatedPublicationId,
-                    Owner = true,
-                    Methodology = new Methodology
-                    {
-                        Versions = ListOf(new MethodologyVersion
-                        {
-                            Status = Draft
-                        }),
-                        OwningPublicationTitle = "Original Title",
-                        OwningPublicationSlug = "original-slug",
-                    }
-                };
-
-                await contentDbContext.PublicationMethodologies.AddRangeAsync(
-                    publicationMethodology,
-                    unrelatedPublicationMethodology);
-                await contentDbContext.SaveChangesAsync();
-            }
-
-            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
-            {
-                var redirectsCacheService = new Mock<IRedirectsCacheService>(MockBehavior.Strict);
-                redirectsCacheService.Setup(mock => mock.UpdateRedirects())
-                    .ReturnsAsync(new RedirectsViewModel(new List<RedirectViewModel>(), new List<RedirectViewModel>()));
-
-                var service = SetupMethodologyService(contentDbContext,
-                    redirectsCacheService: redirectsCacheService.Object);
-                await service.PublicationTitleOrSlugChanged(publicationId, "original-slug", "New Title", "new-slug");
-            }
-
-            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
-            {
-                var publicationMethodology = await contentDbContext
-                    .PublicationMethodologies
-                    .Include(m => m.Methodology)
-                    .SingleAsync(m => m.PublicationId == unrelatedPublicationId);
-
-                // This Methodology was not related to the Publication being updated, and so was not affected by the update.
-                Assert.Equal("Original Title", publicationMethodology.Methodology.OwningPublicationTitle);
-                Assert.Equal("original-slug", publicationMethodology.Methodology.OwningPublicationSlug);
-            }
-        }
-
-        [Fact]
-        public async Task PublicationTitleOrSlugChanged_DoesNotAffectUnownedMethodologies()
-        {
-            var publicationId = Guid.NewGuid();
-
-            var contentDbContextId = Guid.NewGuid().ToString();
-
-            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
-            {
-                var publicationMethodology = new PublicationMethodology
-                {
-                    PublicationId = publicationId,
-                    Owner = false,
-                    Methodology = new Methodology
-                    {
-                        Versions = ListOf(new MethodologyVersion()),
-                        OwningPublicationTitle = "Original Title",
-                        OwningPublicationSlug = "original-slug",
-                    }
-                };
-
-                await contentDbContext.PublicationMethodologies.AddAsync(publicationMethodology);
-                await contentDbContext.SaveChangesAsync();
-            }
-
-            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
-            {
-                var service = SetupMethodologyService(contentDbContext);
-                await service.PublicationTitleOrSlugChanged(publicationId,
-                    "original-slug", "New Title", "new-slug");
-            }
-
-            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
-            {
-                var publicationMethodology = await contentDbContext
-                    .PublicationMethodologies
-                    .Include(m => m.Methodology)
-                    .SingleAsync(m => m.PublicationId == publicationId);
-
-                // This Methodology was not owned by the Publication being updated, and so was not affected by the update.
-                Assert.Equal("Original Title", publicationMethodology.Methodology.OwningPublicationTitle);
-                Assert.Equal("original-slug", publicationMethodology.Methodology.OwningPublicationSlug);
-            }
-        }
-
-        [Fact]
-        public async Task PublicationTitleOrSlugChanged_MethodologySlugHasAlreadyBeenAmended()
-        {
-            var publicationId = Guid.NewGuid();
-            var latestPublishedVersionId = Guid.NewGuid();
-
-            var contentDbContextId = Guid.NewGuid().ToString();
-            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
-            {
-                var publicationMethodology = new PublicationMethodology
-                {
-                    PublicationId = publicationId,
-                    Owner = true,
-                    Methodology = new Methodology
-                    {
-                        LatestPublishedVersionId = latestPublishedVersionId,
-                        OwningPublicationTitle = "Original title",
-                        OwningPublicationSlug = "original-slug",
-                        Versions = ListOf(new MethodologyVersion
-                        {
-                            Id = latestPublishedVersionId,
-                            AlternativeSlug = "alternative-slug",
-                        }),
-                    }
-                };
-
-                await contentDbContext.PublicationMethodologies.AddAsync(publicationMethodology);
-                await contentDbContext.SaveChangesAsync();
-            }
-
-            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
-            {
-                var redirectsCacheService = new Mock<IRedirectsCacheService>(MockBehavior.Strict);
-                redirectsCacheService.Setup(mock => mock.UpdateRedirects())
-                    .ReturnsAsync(new RedirectsViewModel(new List<RedirectViewModel>(), new List<RedirectViewModel>()));
-
-                var service = SetupMethodologyService(contentDbContext,
-                    redirectsCacheService: redirectsCacheService.Object);
-                await service.PublicationTitleOrSlugChanged(publicationId, "original-slug", "New Title", "new-slug");
-            }
-
-            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
-            {
-                var publicationMethodology = await contentDbContext
-                    .PublicationMethodologies
-                    .Include(m => m.Methodology)
-                    .ThenInclude(m => m.Versions)
-                    .SingleAsync(m => m.PublicationId == publicationId);
-
-                Assert.Equal("New Title", publicationMethodology.Methodology.OwningPublicationTitle);
-                Assert.Equal("new-slug", publicationMethodology.Methodology.OwningPublicationSlug);
-
-                // The MethodologyVersion has already had an AlternativeSlug set. It doesn't have a AlternativeTitle
-                // set. So the MethodologyVersion title is updated, but the slug remains the same.
-                Assert.Equal("New Title", publicationMethodology.Methodology.Versions[0].Title);
-                Assert.Equal("alternative-slug", publicationMethodology.Methodology.Versions[0].Slug);
-
-                // No redirect created as slug hasn't changed
-                var methodologyRedirects = await contentDbContext.MethodologyRedirects
-                    .ToListAsync();
-                Assert.Empty(methodologyRedirects);
-            }
-        }
-
-        [Fact]
-        public async Task PublicationTitleOrSlugChanged_MethodologyAlreadyPubliclyAvailable()
-        {
-            var publicationId = Guid.NewGuid();
-            var latestPublishedVersionId = Guid.NewGuid();
-
-            var contentDbContextId = Guid.NewGuid().ToString();
-            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
-            {
-                var publicationMethodology = new PublicationMethodology
-                {
-                    Publication = new Publication
-                    {
-                        Id = publicationId,
-                        LatestPublishedRelease = new Release()
-                    },
-                    Owner = true,
-                    Methodology = new Methodology
-                    {
-                        LatestPublishedVersionId = latestPublishedVersionId,
-                        Versions = ListOf(new MethodologyVersion
-                        {
-                            Id = latestPublishedVersionId,
-                        }),
-                        OwningPublicationTitle = "Original title",
-                        OwningPublicationSlug = "original-slug",
-                    }
-                };
-
-                await contentDbContext.PublicationMethodologies.AddAsync(publicationMethodology);
-                await contentDbContext.SaveChangesAsync();
-            }
-
-            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
-            {
-                var redirectsCacheService = new Mock<IRedirectsCacheService>(MockBehavior.Strict);
-                redirectsCacheService.Setup(mock => mock.UpdateRedirects())
-                    .ReturnsAsync(new RedirectsViewModel(new List<RedirectViewModel>(), new List<RedirectViewModel>()));
-
-                var service = SetupMethodologyService(contentDbContext,
-                    redirectsCacheService: redirectsCacheService.Object);
-                await service.PublicationTitleOrSlugChanged(publicationId, "original-slug", "New Title", "new-slug");
-            }
-
-            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
-            {
-                var publicationMethodology = await contentDbContext
-                    .PublicationMethodologies
-                    .Include(m => m.Methodology)
-                    .SingleAsync(m => m.PublicationId == publicationId);
-
-                Assert.Equal("New Title", publicationMethodology.Methodology.OwningPublicationTitle);
-                Assert.Equal("new-slug", publicationMethodology.Methodology.OwningPublicationSlug);
-
-                var redirect = await contentDbContext
-                    .MethodologyRedirects
-                    .SingleAsync();
-
-                Assert.Equal(latestPublishedVersionId, redirect.MethodologyVersionId);
-                Assert.Equal("original-slug", redirect.Slug);
-            }
-        }
-
-        [Fact]
-        public async Task PublicationTitleOrSlugChanged_NoDuplicateMethodologyRedirectAllowed()
-        {
-            var publicationId = Guid.NewGuid();
-            var latestPublishedVersionId = Guid.NewGuid();
-
-            var contentDbContextId = Guid.NewGuid().ToString();
-            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
-            {
-                var publicationMethodology = new PublicationMethodology
-                {
-                    Publication = new Publication
-                    {
-                        Id = publicationId,
-                        LatestPublishedRelease = new Release()
-                    },
-                    Owner = true,
-                    Methodology = new Methodology
-                    {
-                        LatestPublishedVersionId = latestPublishedVersionId,
-                        Versions = new List<MethodologyVersion>
-                        {
-                            new ()
-                            {
-                                Id = latestPublishedVersionId,
-                                Version = 0,
-                            },
-                        },
-                        OwningPublicationTitle = "Original title",
-                        OwningPublicationSlug = "original-slug",
-                    }
-                };
-
-                var methodologyRedirect = new MethodologyRedirect
-                {
-                    MethodologyVersion = publicationMethodology.Methodology.Versions[0],
-                    // The current methodology slug can match a redirect slug if a user changes via inherited
-                    // publication slug multiple times
-                    Slug = "original-slug",
-                };
-
-                await contentDbContext.PublicationMethodologies.AddAsync(publicationMethodology);
-                await contentDbContext.MethodologyRedirects.AddAsync(methodologyRedirect);
-                await contentDbContext.SaveChangesAsync();
-            }
-
-            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
-            {
-                var redirectsCacheService = new Mock<IRedirectsCacheService>(MockBehavior.Strict);
-                redirectsCacheService.Setup(mock => mock.UpdateRedirects())
-                    .ReturnsAsync(new RedirectsViewModel(new List<RedirectViewModel>(), new List<RedirectViewModel>()));
-
-                var service = SetupMethodologyService(contentDbContext,
-                    redirectsCacheService: redirectsCacheService.Object);
-                await service.PublicationTitleOrSlugChanged(publicationId, "original-slug",
-                    "New Title", "new-slug");
-            }
-
-            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
-            {
-                var publicationMethodology = await contentDbContext
-                    .PublicationMethodologies
-                    .Include(m => m.Methodology.Versions)
-                    .SingleAsync(m => m.PublicationId == publicationId);
-
-                Assert.Equal("New Title", publicationMethodology.Methodology.OwningPublicationTitle);
-                Assert.Equal("new-slug", publicationMethodology.Methodology.OwningPublicationSlug);
-
-                var redirects = await contentDbContext
-                    .MethodologyRedirects
-                    .ToListAsync();
-
-                var redirect = Assert.Single(redirects);
-
-                Assert.Equal(latestPublishedVersionId, redirect.MethodologyVersionId);
-                Assert.Equal("original-slug", redirect.Slug);
-            }
-        }
-
-        [Fact]
-        public async Task PublicationTitleOrSlugChanged_UnpublishedMethodologyAmendmentNeedsRedirectIfInheritedPubSlugChanges()
-        {
-            var publicationId = Guid.NewGuid();
-            var oldVersionId = Guid.NewGuid();
-            var latestPublishedVersionId = Guid.NewGuid();
-            var latestVersionId = Guid.NewGuid();
-
-            var contentDbContextId = Guid.NewGuid().ToString();
-            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
-            {
-                var publicationMethodology = new PublicationMethodology
-                {
-                    Publication = new Publication
-                    {
-                        Id = publicationId,
-                        LatestPublishedRelease = new Release()
-                    },
-                    Owner = true,
-                    Methodology = new Methodology
-                    {
-                        LatestPublishedVersionId = latestPublishedVersionId,
-                        Versions = new List<MethodologyVersion>
-                        {
-                            new ()
-                            {
-                                Id = latestPublishedVersionId,
-                                Version = 0,
-                            },
-                            new ()
-                            {
-                                Id = latestVersionId,
-                                Version = 1,
-                                AlternativeSlug = "methodology-alternative-slug",
-                                PreviousVersionId = latestPublishedVersionId,
-                            }
-                        },
-                        OwningPublicationTitle = "Current title",
-                        OwningPublicationSlug = "current-slug",
-                    }
-                };
-
-                var methodologyRedirect = new MethodologyRedirect
-                {
-                    // This would have been created when the unpublished amendment's AlternativeSlug was set.
-                    MethodologyVersionId = latestVersionId,
-                    Slug = "current-slug",
-                };
-
-                await contentDbContext.PublicationMethodologies.AddAsync(publicationMethodology);
-                await contentDbContext.MethodologyRedirects.AddAsync(methodologyRedirect);
-                await contentDbContext.SaveChangesAsync();
-            }
-
-            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
-            {
-                var redirectsCacheService = new Mock<IRedirectsCacheService>(MockBehavior.Strict);
-                redirectsCacheService.Setup(mock => mock.UpdateRedirects())
-                    .ReturnsAsync(new RedirectsViewModel(new List<RedirectViewModel>(), new List<RedirectViewModel>()));
-
-                var service = SetupMethodologyService(contentDbContext,
-                    redirectsCacheService: redirectsCacheService.Object);
-                await service.PublicationTitleOrSlugChanged(publicationId, "current-slug",
-                    "New Title", "new-slug");
-            }
-
-            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
-            {
-                var publicationMethodology = await contentDbContext
-                    .PublicationMethodologies
-                    .Include(m => m.Methodology.Versions)
-                    .SingleAsync(m => m.PublicationId == publicationId);
-
-                Assert.Equal("New Title", publicationMethodology.Methodology.OwningPublicationTitle);
-                Assert.Equal("new-slug", publicationMethodology.Methodology.OwningPublicationSlug);
-
-                var redirects = await contentDbContext
-                    .MethodologyRedirects
-                    .ToListAsync();
-
-                Assert.Equal(2, redirects.Count);
-
-                // "current-slug" redirect for latestVersion is removed, as a "current-slug" redirect has been added
-                // for the latestPublishedVersion
-
-                Assert.Equal(latestPublishedVersionId, redirects[0].MethodologyVersionId);
-                Assert.Equal("current-slug", redirects[0].Slug);
-
-                // A new redirect for the unpublished amendment, as otherwise we have no redirect for
-                // "new-slug" once the amendment is published
-                Assert.Equal(latestVersionId, redirects[1].MethodologyVersionId);
-                Assert.Equal("new-slug", redirects[1].Slug);
-            }
-        }
-
         private static MethodologyService SetupMethodologyService(
             ContentDbContext contentDbContext,
             IPersistenceHelper<ContentDbContext>? contentPersistenceHelper = null,
-            IMethodologyVersionRepository? methodologyVersionRepository = null,
-            IRedirectsCacheService? redirectsCacheService = null)
+            IMethodologyVersionRepository? methodologyVersionRepository = null)
         {
             return new(
                 contentDbContext,
                 contentPersistenceHelper ?? new PersistenceHelper<ContentDbContext>(contentDbContext),
                 MapperUtils.ContentMapper(contentDbContext),
-                methodologyVersionRepository ?? Mock.Of<IMethodologyVersionRepository>(MockBehavior.Strict),
-                redirectsCacheService ?? Mock.Of<IRedirectsCacheService>(MockBehavior.Strict)
+                methodologyVersionRepository ?? Mock.Of<IMethodologyVersionRepository>(MockBehavior.Strict)
             );
         }
     }
