+++ src/GovUk.Education.ExploreEducationStatistics.Content.Services/MethodologyService.cs
@@ -11,7 +11,6 @@ using GovUk.Education.ExploreEducationStatistics.Content.Model;
 using GovUk.Education.ExploreEducationStatistics.Content.Model.Database;
 using GovUk.Education.ExploreEducationStatistics.Content.Model.Repository.Interfaces;
 using GovUk.Education.ExploreEducationStatistics.Content.Services.Interfaces;
-using GovUk.Education.ExploreEducationStatistics.Content.Services.Interfaces.Cache;
 using GovUk.Education.ExploreEducationStatistics.Content.Services.ViewModels;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -24,19 +23,16 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Services
         private readonly IPersistenceHelper<ContentDbContext> _persistenceHelper;
         private readonly IMapper _mapper;
         private readonly IMethodologyVersionRepository _methodologyVersionRepository;
-        private readonly IRedirectsCacheService _redirectsCacheService;
 
         public MethodologyService(ContentDbContext contentDbContext,
             IPersistenceHelper<ContentDbContext> persistenceHelper,
             IMapper mapper,
-            IMethodologyVersionRepository methodologyVersionRepository,
-            IRedirectsCacheService redirectsCacheService)
+            IMethodologyVersionRepository methodologyVersionRepository)
         {
             _contentDbContext = contentDbContext;
             _persistenceHelper = persistenceHelper;
             _mapper = mapper;
             _methodologyVersionRepository = methodologyVersionRepository;
-            _redirectsCacheService = redirectsCacheService;
         }
 
         public async Task<Either<ActionResult, MethodologyVersionViewModel>> GetLatestMethodologyBySlug(string slug)
@@ -136,99 +132,6 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Services
             var latestPublishedMethodologies =
                 await _methodologyVersionRepository.GetLatestPublishedVersionByPublication(publicationId);
             return _mapper.Map<List<MethodologyVersionSummaryViewModel>>(latestPublishedMethodologies);
-        }
-
-        // This method is responsible for keeping Methodology Titles and Slugs in sync with their owning Publications
-        // where appropriate.  Methodologies always keep track of their owning Publication's titles and slugs for
-        // optimisation purposes.
-        public async Task PublicationTitleOrSlugChanged(Guid publicationId, string originalSlug, string updatedTitle,
-            string updatedSlug)
-        {
-            var slugChanged = originalSlug != updatedSlug;
-
-            var ownedMethodology = await _contentDbContext
-                .PublicationMethodologies
-                .Include(pm => pm.Methodology.LatestPublishedVersion)
-                .Include(pm => pm.Methodology.Versions)
-                .Where(pm => pm.PublicationId == publicationId && pm.Owner)
-                .Select(pm => pm.Methodology)
-                .SingleOrDefaultAsync();
-
-            if (ownedMethodology == null)
-            {
-                return;
-            }
-
-            ownedMethodology.OwningPublicationTitle = updatedTitle;
-            ownedMethodology.OwningPublicationSlug = updatedSlug;
-
-            _contentDbContext.Methodologies.Update(ownedMethodology);
-
-            if (slugChanged)
-            {
-                // A redirect is (with one exception, below) is only needed for the LatestPublishedVersion.
-                // An unpublished amendment doesn't need a redirect because:
-                // - if it uses OwningPublicationSlug, it is covered by the LatestPublishedVersion redirect created here
-                // - if it uses AlternativeSlug, a redirect would have been created at the time the AlternativeSlug
-                //   was set (and that redirect will become active when that version is published).
-                if (ownedMethodology.LatestPublishedVersion is { AlternativeSlug: null }
-                    // guard against duplicates due to users making multiple publication slug changes
-                    && !await _contentDbContext.MethodologyRedirects
-                        .AnyAsync(mr =>
-                            mr.MethodologyVersionId == ownedMethodology.LatestPublishedVersionId
-                            && mr.Slug == originalSlug))
-                {
-                    var redirect = new MethodologyRedirect
-                    {
-                        MethodologyVersion = ownedMethodology.LatestPublishedVersion,
-                        Slug = originalSlug,
-                    };
-                    _contentDbContext.MethodologyRedirects.Add(redirect);
-
-                    // This is the exception mentioned above. If we have an unpublished amendment with an
-                    // AlternativeSlug set, and the LatestPublishedVersion inherits the publication slug,
-                    // then we need to create a redirect for the unpublished amendment for the new slug.
-                    if (ownedMethodology.LatestVersion().Id != ownedMethodology.LatestPublishedVersionId
-                        // && ownedMethodology.LatestPublishedVersion is { AlternativeSlug: null} // already checked above
-                        && ownedMethodology.LatestVersion().AlternativeSlug != null)
-                        // && (redirectSlugs.Contains(updatedSlug) && ownedVersionIds.Contains(
-                    {
-                        var unpublishedAmendmentRedirect = new MethodologyRedirect
-                        {
-                            MethodologyVersion = ownedMethodology.LatestVersion(),
-                            Slug = updatedSlug,
-                        };
-                        _contentDbContext.MethodologyRedirects.Add(unpublishedAmendmentRedirect);
-                    }
-
-                    // It's possible we now have two redirects from the same slug. This happens if:
-                    // - An unpublished amendment sets an AlternativeSlug
-                    // - Then the OwningPublicationSlug changes when the LatestPublishedVersion is
-                    //   inheriting it.
-                    // We must delete the unpublished amendment's redirect, as otherwise if the unpublished version
-                    // changes its slug again, no redirect will be created for the methodology's new slug. (If updating
-                    // an unpublished version's slug multiple times, a redirect is only created the first time - see
-                    // MethodologyUpdate code). If this doesn't make any sense, see the unit test
-                    // PublicationTitleOrSlugChanged_NewMethodologyRedirectSlugMatchesExistingRedirectSlug
-                    var redirectToRemove = await _contentDbContext.MethodologyRedirects
-                        .Where(mr =>
-                            mr.MethodologyVersionId == ownedMethodology.LatestVersion().Id
-                            && mr.Slug == originalSlug)
-                        .SingleOrDefaultAsync();
-
-                    if (redirectToRemove != null)
-                    {
-                        _contentDbContext.MethodologyRedirects.Remove(redirectToRemove);
-                    }
-                }
-            }
-
-            await _contentDbContext.SaveChangesAsync();
-
-            if (slugChanged)
-            {
-                await _redirectsCacheService.UpdateRedirects();
-            }
         }
     }
 }
